Add files via upload
The main file to run is i-SPin_main.py. It uses three function modules: (1)  "kdkfunc.py" which performs the main (drift-kick-drift) steps of the algorithm; (2) "conservefunc.py" which tracks the fractional changes in total mass and spin in the system; (3) "rtplotfunc.py" which is a module to plot different things. The default setting is to plot mass and spin density projections along with total mass and spin fractional changes.